# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech talks",
        "description": "API documentation for tech talks",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/user/my-talks": {
            "get": {
                "tags": [
                    "talks"
                ],
                "summary": "List user talks",
                "operationId": "userTalks",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Talk"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Talk": {
                "title": "Talk",
                "description": "Talk",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID of the talk",
                        "type": "number"
                    },
                    "topic": {
                        "title": "topic",
                        "description": "Topic of the talk",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the talk",
                        "type": "string"
                    },
                    "date": {
                        "title": "date",
                        "description": "Date",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "enum": [
                            "published",
                            "draft"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}